@@ -1,5 +1,5 @@
 {
-  "name": "IDFactory",
+  "name": "PaPrint",
   "version": "1.0.0",
   "description": "Tutorial",
   "author": "kennCK <kennettecanales@gmail.com>",
@@ -15,32 +15,38 @@
   "dependencies": {
     "axios": "^0.17.1",
     "axios-fileupload": "0.0.1",
-    "bootstrap": "^3.3.7",
-    "bootstrap-vue": "^2.0.0-rc.11",
-    "chart.js": "^2.7.2",
+    "bootstrap": "^3.4.1",
+    "bootstrap-vue": "^2.0.0-rc.25",
+    "chart.js": "^2.8.0",
     "child_process": "^1.0.2",
     "easy-countries-list": "^1.1.0",
     "font-awesome": "^4.7.0",
     "fs": "0.0.1-security",
     "googleapis": "^27.0.0",
-    "howler": "^2.0.12",
-    "jquery": "^3.2.1",
-    "laravel-echo": "^1.5.3",
+    "howler": "^2.1.2",
+    "jquery": "^3.4.1",
+    "laravel-echo": "^1.5.4",
+    "node": "^12.5.0",
     "popper.js": "^1.14.3",
     "pusher-js": "^4.4.0",
-    "react-player": "^1.6.4",
+    "react-player": "^1.11.1",
     "readline": "^1.3.0",
-    "tether": "^1.4.4",
-    "tooltip.js": "^1.2.0",
-    "vue-chartjs": "^3.3.2",
+    "sass": "^1.22.2",
+    "style-resources-loader": "^1.2.1",
+    "tether": "^1.4.6",
+    "tooltip.js": "^1.3.2",
+    "vue-chartjs": "^3.4.2",
     "vue-clip": "^1.0.0",
+    "vue-color": "^2.7.0",
     "vue-directive-tooltip": "^1.4.5",
+    "vue-loader": "^11.3.4",
     "vue-paypal-checkout": "^3.2.0",
     "vue-router": "^2.8.1",
     "vue-simple-audio": "^1.0.6",
-    "vue-stripe-elements-plus": "^0.2.9",
-    "waveplayer": "^1.0.3",
-    "wavesurfer.js": "^2.0.6",
+    "vue-stripe-elements-plus": "^0.2.10",
+    "vuedraggable": "^2.23.0",
+    "waveplayer": "^1.2.2",
+    "wavesurfer.js": "^2.2.1",
     "web-audio-player": "^1.3.3"
   },
   "devDependencies": {
@@ -56,9 +62,9 @@
     "babel-register": "^6.22.0",
     "chai": "^3.5.0",
     "chalk": "^1.1.3",
-    "chromedriver": "^2.40.0",
-    "connect-history-api-fallback": "^1.3.0",
-    "copy-webpack-plugin": "^4.0.1",
+    "chromedriver": "^2.46.0",
+    "connect-history-api-fallback": "^1.6.0",
+    "copy-webpack-plugin": "^4.6.0",
     "cross-env": "^3.1.4",
     "cross-spawn": "^5.0.1",
     "css-loader": "^0.26.1",
@@ -70,7 +76,7 @@
     "eslint-plugin-promise": "^3.8.0",
     "eslint-plugin-standard": "^2.0.1",
     "eventsource-polyfill": "^0.9.6",
-    "express": "^4.16.1",
+    "express": "^4.17.1",
     "extract-text-webpack-plugin": "^2.1.2",
     "file-loader": "^0.10.0",
     "friendly-errors-webpack-plugin": "^1.7.0",
@@ -88,25 +94,27 @@
     "karma-webpack": "^2.0.2",
     "lolex": "^1.5.2",
     "mocha": "^5.2.0",
-    "nightwatch": "^1.0.6",
+    "nightwatch": "^1.1.13",
+    "node-sass": "^4.12.0",
     "opn": "^4.0.2",
     "optimize-css-assets-webpack-plugin": "^1.3.0",
     "ora": "^1.1.0",
     "phantomjs-prebuilt": "^2.1.14",
-    "rimraf": "^2.6.0",
-    "selenium-server": "^3.12.0",
-    "semver": "^5.3.0",
+    "rimraf": "^2.6.3",
+    "sass-loader": "^7.1.0",
+    "selenium-server": "^3.141.59",
+    "semver": "^5.7.0",
     "sinon": "^1.17.7",
     "sinon-chai": "^2.8.0",
+    "style-loader": "^0.23.1",
     "url-loader": "^1.0.1",
-    "vue": "^2.5.13",
-    "vue-loader": "^11.0.0",
+    "vue": "^2.6.10",
     "vue-style-loader": "^2.0.0",
-    "vue-template-compiler": "^2.1.10",
+    "vue-template-compiler": "^2.6.10",
     "webpack": "^2.7.0",
     "webpack-bundle-analyzer": "^2.13.1",
     "webpack-dev-middleware": "^1.12.0",
-    "webpack-hot-middleware": "^2.22.2",
+    "webpack-hot-middleware": "^2.25.0",
     "webpack-merge": "^2.6.1"
   },
   "engines": {
