@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use App\Template;
 use App\Guide;
 use App\ActiveTemplate;
+use Carbon\Carbon;
 class TemplateController extends APIController
 {
   function __construct(){
@@ -15,67 +16,53 @@ class TemplateController extends APIController
     );
   }
 
+  public function create(Request $request){
+    $data = $request->all();
+    $data['code'] = $this->generateCode();
+    $this->model = new Template();
+    $this->insertDB($data);
+    return $this->response();
+  }
+
   public function retrieve(Request $request){
     $data = $request->all();
-    $active = intval($data['active']);
     $this->model = new Template();
     $this->retrieveDB($data);
     $result = $this->response['data'];
-    if(sizeof($result) > 0){
-      $i = 0;
-      foreach ($result as $key) {
-        $this->response['data'][$i]['active_templates'] = $this->getActiveTemplate($result[$i]['settings'], $result[$i]['id'], $data['account_id']);
-        $this->response['data'][$i]['objects'] = app('App\Http\Controllers\ObjectController')->getObjects($result[$i]['id']);
-        $this->response['data'][$i]['guide'] = ($result[$i]['status'] == 'purchased') ? $this->getGuide($result[$i]['purchased']) : $this->getGuide($result[$i]['id']);
-        if($active === $i){
-          $this->response['data'][$i]['active'] = true;
-        }else{
-          $this->response['data'][$i]['active'] = false;
-        }
-       $i++; 
-      }
-    }
+    // if(sizeof($result) > 0){
+    //   $i = 0;
+    //   foreach ($result as $key) {
+    //     $this->response['data'][$i]['contents'] = response()->json()
+    //     $i++;
+    //   }
+    // }
     return $this->response();
   }
 
-  public function retrieveById($id){
-    $result = Template::where('id', '=', $id)->get();
-    if(sizeof($result) > 0){
-      $i = 0;
-      foreach ($result as $key) {
-        $result[$i]['objects'] = app('App\Http\Controllers\ObjectController')->getObjects($result[$i]['id']);
-       $i++; 
-      }
-    }
-    return (sizeof($result) > 0) ? $result[0] : null;
-  }
-
-  public function getActiveTemplate($settings, $id, $accountId){
-    $result = null;
-    if($settings == 'front'){
-      $result = ActiveTemplate::where('account_id', '=', $accountId)->where('front', '=', $id)->get();
-    }else{
-      $result = ActiveTemplate::where('account_id', '=', $accountId)->where('back', '=', $id)->get();
-    }
-
-    return (sizeof($result) > 0) ? $result[0] : null;
-          
-  }
-
-  public function retriveTemplateOnly(Request $request){
+  public function retrieveCategories(Request $request){
     $data = $request->all();
-    $this->model = new Template();
-    $this->retrieveDB($data);
-    return $this->response();
+    $result = Template::where('account_id', '=', $data['account_id'])->orderBy('category', $data['sort'])->get(['category'])->groupBy('category');
+
+    $keys = array();
+
+    foreach ($result as $key) {
+      $keys[] = $key[0]['category'];
+    }
+    return response()->json(array(
+      'data' => $keys,
+      'error' => null,
+      'timestamps' => Carbon::now()
+    ));
+
   }
 
-  public function getGuide($id){
-    $result = Guide::where('payload_value', '=', $id)->where('payload', '=', 'template')->get();
-    return (sizeof($result) > 0) ? $result[0] : null;
-  }
-
-  public function getTemplateDetails($templateId){
-    $result = Template::where('id', '=', $templateId)->get();
-    return (sizeof($result) > 0) ? $result[0] : null;     
+  public function generateCode(){
+    $code = substr(str_shuffle("0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ"), 0, 64);
+    $codeExist = Template::where('code', '=', $code)->get();
+    if(sizeof($codeExist) > 0){
+      $this->generateCodee();
+    }else{
+      return $code;
+    }
   }
 }
