@@ -7,5 +7,5 @@ use Illuminate\Database\Eloquent\Model;
 class Template extends APIModel
 {
     protected $table = 'templates';
-    protected $fillable = ['account_id', 'title', 'settings', 'width', 'height', 'status', 'categories', 'price'];
+    protected $fillable = ['code', 'account_id', 'title', 'contents', 'status', 'categories'];
 }
