@@ -177,11 +177,11 @@ return [
         App\Providers\BroadcastServiceProvider::class,
         App\Providers\EventServiceProvider::class,
         App\Providers\RouteServiceProvider::class,
-        LaravelQRCode\Providers\QRCodeServiceProvider::class,
         Increment\Account\AccountServiceProvider::class,
         Increment\Common\Comment\CommentServiceProvider::class,
         Increment\Common\Rating\RatingServiceProvider::class,
         Increment\Common\Image\ImageServiceProvider::class,
+        Increment\Common\Tutorial\TutorialServiceProvider::class,
         Increment\Marketplace\MarketplaceServiceProvider::class,
         Increment\Messenger\MessengerServiceProvider::class,
         Increment\Plan\PlanServiceProvider::class,
