@@ -4,20 +4,20 @@
             <div class="container-fluid">
                 <div class="row">
                     <div class="col-lg-12">
-                    <!-- <i class="fa fa-copyright" aria-hidden="true"></i>  -->
-                    <!-- <b>Talk Fluent Spanish</b> 2018 -->
+                      A product of <a href="http://increment.ltd" target="_BLANK">Increment Technologies</a>
                     </div>
                 </div>
             </div>
         </div>
     </footer>
 </template>
-<style type="text/css">
+<style lang="scss">
+@import "~assets/style/colors.scss";
 footer {
   width:100%;
   height:30px;   /* Height of the footer */
   float: left;
-  background: #fff;
+  background: $white;
   padding: 10px 0 10px 0;
   font-size: 12px;
 }
