@@ -1,10 +1,10 @@
 <template>
 	<div class="body">
-		<landing-banner></landing-banner>
     <landing-steps></landing-steps>
+    <landing-editor></landing-editor>
     <landing-bizmarketplace></landing-bizmarketplace>
     <landing-mc></landing-mc>
-    <landing-features></landing-features>
+    <landing-filemanagement></landing-filemanagement>
 	</div>
 </template>
 <script>
@@ -12,11 +12,12 @@ import ROUTER from '../../router'
 import AUTH from '../../services/auth'
 export default {
   components: {
-    'landing-banner': require('modules/home/Landing/Banner.vue'),
+    'landing-editor': require('modules/home/Landing/Editor.vue'),
     'landing-bizmarketplace': require('modules/home/Landing/BizMarketplace.vue'),
     'landing-steps': require('modules/home/Landing/Steps.vue'),
     'landing-features': require('modules/home/Landing/Features.vue'),
-    'landing-mc': require('modules/home/Landing/MessageAndComment.vue')
+    'landing-mc': require('modules/home/Landing/MessageAndComment.vue'),
+    'landing-filemanagement': require('modules/home/Landing/FileManagement.vue')
   },
   mounted(){
   },
