@@ -3,28 +3,23 @@ let CONFIG = require('config.js')
 let beforeEnter = (to, from, next) => {
   // TODO Redirect if no token when token is required in meta.tokenRequired
   AUTH.currentPath = to.path
-  let userID = localStorage.getItem('account_id')
+  let userID = parseInt(localStorage.getItem('account_id'))
   let token = localStorage.getItem('usertoken')
-  if(userID > 0 || token !== null || to.meta.tokenRequired !== true){
-    if(to.path === '/' && userID > 0 && token !== null){
-      next({path: '/'})
+  if(token !== null && userID > 0){
+    if(to.path === '/' || to.path === '/'){
+      next({path: '/templates'})
     }else{
       next()
     }
-  }else{
+  }else if(to.meta.tokenRequired === true){
     next({path: '/'})
-    // if(userID <= 0){
-    //   next({
-    //     path: '/'
-    //   })
-    // }else{
-    //   next()
-    // }
+  }else{
+    next()
   }
 }
 var devRoutes = []
-let canales = require('./dev_routes/canales.js')
-devRoutes = devRoutes.concat(canales.default.routes)
+let paprint = require('./dev_routes/paprint.js')
+devRoutes = devRoutes.concat(paprint.default.routes)
 for(let x = 0; x < devRoutes.length; x++){
   devRoutes[x]['beforeEnter'] = beforeEnter
 }
